Infographics of Best Streaming Services
Over the past few years now, the world has evolved from traditional movie suggestions by word of
mouth, advert posts on billboards, radio jingles, to a more advanced and modern way of
communicating quality information that tends to inform, educate, entertain, or even compel, through
digital means for the good interest of either its targeted audience, publishers, or both.

This could not have been possible without the dependency of historical datasets (data) to make
powerful assumptions as we now have in Machine Learning and AI enabled applications; or to tell
stories of what has happened in a business transaction of a certain period for better decision making
in future; or to make compelling discoveries and suggestions based on the personal choices we make
in our daily lives shopping, movies, vacations, etc.

In this project, we have used infographic designs and analytics to create a pathway of discovering to
the movie lovers, the best streaming channel between Amazon, Disney Plus, Hulu, and Netflix
that suits the most for different reasons. We make the discoveries while you decide what is best for you.